+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.1.0] - Unreleased
+## [0.1.0] - 2026-05-21
 
 Initial release.